+++ q_compress/src/wrapped/decompressor.rs
@@ -1,9 +1,18 @@
 use std::io::Write;
-use crate::base_decompressor::{BaseDecompressor, Step};
+
 use crate::{ChunkMetadata, DecompressorConfig, Flags};
+use crate::base_decompressor::{BaseDecompressor, Step};
+use crate::bit_words::BitWords;
 use crate::data_types::NumberLike;
 use crate::errors::QCompressResult;
 
+/// Converts wrapped Quantile-compressed data into [`Flags`],
+/// [`ChunkMetadata`], and vectors of numbers.
+///
+/// All decompressor methods leave its state unchanged if they return an
+/// error.
+///
+/// You can use the decompressor at a data page level.
 #[derive(Clone, Debug, Default)]
 pub struct Decompressor<T: NumberLike>(BaseDecompressor<T>);
 
@@ -14,7 +23,7 @@ impl<T: NumberLike> Decompressor<T> {
   }
 
   /// Reads the header, returning its [`Flags`] and updating this
-  /// `Decompressor`'s state.
+  /// decompressor's state.
   /// Will return an error if this decompressor has already parsed a header,
   /// is not byte-aligned,
   /// runs out of data,
@@ -24,19 +33,74 @@ impl<T: NumberLike> Decompressor<T> {
     self.0.header(true)
   }
 
-  /// TODO
-  pub fn chunk_metadata(&mut self) -> QCompressResult<Option<ChunkMetadata<T>>> {
-    self.0.state.check_step_among(&[Step::StartOfChunk, Step::StartOfDataPage], "read chunk metadata")?;
-    self.0.state.body_decompressor = None;
+  /// Reads the chunk metadata, returning its metadata and updating the
+  /// decompressor's state.
+  /// Will return an error if the decompressor has not parsed the header,
+  /// runs out of data,
+  /// or finds any corruptions.
+  ///
+  /// This can be used regardless of whether the decompressor has finished
+  /// reading all data pages from the preceding chunk.
+  pub fn chunk_metadata(&mut self) -> QCompressResult<ChunkMetadata<T>> {
+    self.0.state.check_step_among(
+      &[Step::StartOfChunk, Step::StartOfDataPage, Step::MidDataPage],
+      "read chunk metadata",
+    )?;
 
-    self.0.chunk_metadata_internal()
+    self.0.with_reader(|reader, state, _| {
+      let meta = ChunkMetadata::<T>::parse_from(reader, state.flags.as_ref().unwrap())?;
+
+      state.chunk_meta = Some(meta.clone());
+      state.body_decompressor = None;
+      Ok(meta)
+    })
   }
-  // TODO
+  
+  /// Initializes the decompressor for the next data page, reading in the
+  /// data page's metadata but not the compressed body.
+  /// Will return an error if the decompressor is not in the middle of a
+  /// chunk, runs out of data, or finds any corruptions. 
+  ///
+  /// This can be used regardless of whether the decompressor has finished
+  /// reading the previous data page.
+  pub fn begin_data_page(
+    &mut self,
+    n: usize,
+    compressed_page_size: usize,
+  ) -> QCompressResult<()> {
+    self.0.state.check_step_among(&[Step::StartOfDataPage, Step::MidDataPage], "begin data page")?;
+    self.0.with_reader(|reader, state, _| {
+      state.body_decompressor = Some(state.new_body_decompressor(reader, n, compressed_page_size)?);
+      Ok(())
+    })
+  }
+
+  /// TODO
+  pub fn next_batch(
+    &mut self,
+    limit: usize,
+  ) -> QCompressResult<Vec<T>> {
+    self.0.state.check_step(Step::MidDataPage, "read next batch")?;
+    self.0.with_reader(|reader, state, _| {
+      let bd = state.body_decompressor.as_mut().unwrap();
+      let numbers = bd.decompress_next_batch(reader, limit, true)?;
+      if numbers.finished_body {
+        state.body_decompressor = None;
+      }
+      Ok(numbers.nums)
+    })
+  }
+
+  /// Reads an entire data page, returning its numbers.
   pub fn data_page(
     &mut self,
     n: usize,
     compressed_page_size: usize,
   ) -> QCompressResult<Vec<T>> {
+    self.0.state.check_step_among(
+      &[Step::StartOfDataPage, Step::MidDataPage],
+      "data page",
+    )?;
     self.0.data_page_internal(n, compressed_page_size)
   }
 
@@ -45,6 +109,16 @@ impl<T: NumberLike> Decompressor<T> {
   /// Note that calling this too frequently can cause performance issues.
   pub fn free_compressed_memory(&mut self) {
     self.0.free_compressed_memory()
+  }
+  
+  /// Clears any data written to the decompressor but not yet decompressed.
+  /// As an example, if you want to want to read the first 5 numbers from each
+  /// data page, you might write each compressed data page to the decompressor,
+  /// then repeatedly call [`begin_data_page`], [`next_nums`], and
+  /// this method.
+  pub fn clear_compressed_bytes(&mut self) {
+    self.0.words = BitWords::default();
+    self.0.state.bit_idx = 0;
   }
 
   /// Returns the current bit position into the compressed data the