+++ q_compress/src/wrapped/mod.rs
@@ -1,5 +1,6 @@
 pub use compressor::Compressor;
 pub use decompressor::Decompressor;
+pub use crate::chunk_spec::ChunkSpec;
 
 mod compressor;
 mod decompressor;